segocr/generator/font_manager.py
"""Font loading, validation, caching, and weighted sampling.

Implements the Font Manager spec from Implementation Guide §3.1.
"""
from __future__ import annotations

from pathlib import Path
from typing import TYPE_CHECKING

if TYPE_CHECKING:
    from PIL.ImageFont import FreeTypeFont


class FontManager:
    """Manages the font library for text rendering.

    Walks the font root directory, validates each .ttf/.otf by attempting to
    load and render every character in the active character set, discards
    fonts that crash or have missing glyphs, and caches the validation
    results to disk so subsequent runs skip the expensive scan.
    """

    def __init__(self, config: dict) -> None:
        """
        Args:
            config: the ``fonts`` section of the YAML config.
        """
        self.config = config
        self.root_dir = Path(config["root_dir"])
        self.cache_path = Path(config["cache_path"])
        self.fonts_by_category: dict[str, list[Path]] = {}
        self.bbox_cache: dict[tuple[Path, str], tuple[int, int, int, int]] = {}
        raise NotImplementedError("FontManager.__init__ — Week 2")

    def sample_font(self) -> tuple["FreeTypeFont", str]:
        """Sample a random font weighted by category distribution.

        Returns:
            (font_object, font_category)
        """
        raise NotImplementedError("FontManager.sample_font — Week 2")

    def get_char_bbox(
        self, font: "FreeTypeFont", char: str, size: int
    ) -> tuple[int, int, int, int]:
        """Get pre-computed bounding box for a character at given size.

        Bounding boxes are pre-computed at a reference size (48px) and
        scaled proportionally for other sizes.
        """
        raise NotImplementedError("FontManager.get_char_bbox — Week 2")

    def get_all_fonts(self) -> list["FreeTypeFont"]:
        """Return all validated fonts."""
        raise NotImplementedError("FontManager.get_all_fonts — Week 2")

    def _validate_font(self, font_path: Path) -> bool:
        """Try to load and render the full character set with this font.

        Wraps every operation in try/except — some .ttf files crash Pillow.
        """
        raise NotImplementedError("FontManager._validate_font — Week 2")

    def _load_or_build_cache(self) -> None:
        """Read cache from disk; if missing or invalid, re-scan and rebuild."""
        raise NotImplementedError("FontManager._load_or_build_cache — Week 2")
